Construccion del crud colores

+++ src/main/java/com/multiplicatalent/talent/repositories/IColoresRepository.java
@@ -3,7 +3,7 @@ package com.multiplicatalent.talent.repositories;
 import com.multiplicatalent.talent.models.Color;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IColoresRepository extends JpaRepository<Color, Integer> {
+public interface IColoresRepository extends JpaRepository<Color, Long> {
 
 
 }
